refactor(backend): architecture redesign with skinny controllers
- Centralized service interfaces in NikoNiko.Core/Interfaces
- Created SprintService and MoodService to encapsulate business logic
- Refactored TeamsController, SprintsController, MoodEntriesController, and UsersController to delegate logic to services
- Updated DI registration in Program.cs
- Verified system stability with full integration test suite
