+++ api/NikoNiko.Api/Controllers/SprintsController.cs
@@ -1,12 +1,8 @@
 using System.Security.Claims;
-
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-
 using NikoNiko.Core.DTOs.Sprint;
-using NikoNiko.Core.Models;
-using NikoNiko.Data;
+using NikoNiko.Core.Interfaces;
 
 namespace NikoNiko.Api.Controllers;
 
@@ -18,19 +14,16 @@ namespace NikoNiko.Api.Controllers;
 [Route("api/[controller]")]
 public class SprintsController : ControllerBase
 {
-    private readonly ApplicationDbContext _context;
+    private readonly ISprintService _sprintService;
 
-    public SprintsController(ApplicationDbContext context)
+    public SprintsController(ISprintService sprintService)
     {
-        _context = context;
+        _sprintService = sprintService;
     }
 
     /// <summary>
     /// Gets a list of all sprints, optionally filtered by teamId.
-    /// A regular user can only see sprints for teams they are a member of.
-    /// A super-admin can see all sprints.
     /// </summary>
-    /// <returns>A list of SprintDto objects.</returns>
     [HttpGet]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status401Unauthorized)]
@@ -43,59 +36,21 @@ public class SprintsController : ControllerBase
         }
 
         var isSuperAdmin = User.HasClaim("is_super_admin", "true");
-
-        var query = _context.Sprints.AsQueryable();
-
-        if (teamId.HasValue)
-        {
-            query = query.Where(s => s.TeamId == teamId.Value);
-        }
-
-        if (!isSuperAdmin)
-        {
-            query = query.Where(s => s.Team.AdminId == userId || s.Team.TeamUsers.Any(tu => tu.UserId == userId));
-        }
-
-        var sprints = await query
-            .Select(s => new SprintDto
-            {
-                Id = s.Id,
-                Name = s.Name,
-                StartDate = s.StartDate,
-                EndDate = s.EndDate,
-                TeamId = s.TeamId
-            })
-            .ToListAsync();
-
+        var sprints = await _sprintService.GetSprintsAsync(teamId, userId, isSuperAdmin);
         return Ok(sprints);
     }
 
     /// <summary>
     /// Gets a specific sprint by its ID.
-    /// A user must be a member of the sprint's team, or a super-admin.
     /// </summary>
-    /// <param name="sprintId">The ID of the sprint.</param>
-    /// <returns>The SprintDto object, or NotFound if the sprint does not exist.</returns>
     [HttpGet("{sprintId}")]
     [Authorize(Policy = "IsTeamMember")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<ActionResult<SprintDto>> GetSprint(Guid sprintId)
     {
-        var sprint = await _context.Sprints
-            .Where(s => s.Id == sprintId)
-            .Select(s => new SprintDto
-            {
-                Id = s.Id,
-                Name = s.Name,
-                StartDate = s.StartDate,
-                EndDate = s.EndDate,
-                TeamId = s.TeamId
-            })
-            .FirstOrDefaultAsync();
-
+        var sprint = await _sprintService.GetSprintByIdAsync(sprintId);
         if (sprint == null)
         {
             return NotFound();
@@ -106,10 +61,7 @@ public class SprintsController : ControllerBase
 
     /// <summary>
     /// Creates a new sprint.
-    /// Only a team-admin or super-admin can create a sprint.
     /// </summary>
-    /// <param name="createSprintDto">The data needed to create a sprint.</param>
-    /// <returns>The SprintDto of the created sprint.</returns>
     [HttpPost]
     [ProducesResponseType(StatusCodes.Status201Created)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
@@ -123,139 +75,91 @@ public class SprintsController : ControllerBase
             return Unauthorized("User ID not found or invalid.");
         }
 
-        var team = await _context.Teams.FindAsync(createSprintDto.TeamId);
-        if (team == null)
+        var isSuperAdmin = User.HasClaim("is_super_admin", "true");
+
+        try
         {
-            ModelState.AddModelError(nameof(createSprintDto.TeamId), "Team not found.");
+            var sprintDto = await _sprintService.CreateSprintAsync(createSprintDto, userId, isSuperAdmin);
+            return CreatedAtAction(nameof(GetSprint), new { sprintId = sprintDto.Id }, sprintDto);
+        }
+        catch (KeyNotFoundException ex)
+        {
+            ModelState.AddModelError(nameof(createSprintDto.TeamId), ex.Message);
             return BadRequest(ModelState);
         }
-
-        var isSuperAdmin = User.HasClaim("is_super_admin", "true");
-        var isTeamAdmin = team.AdminId == userId;
-
-        if (!isSuperAdmin && !isTeamAdmin)
+        catch (UnauthorizedAccessException)
         {
             return Forbid();
         }
-
-
-        if (createSprintDto.EndDate <= createSprintDto.StartDate)
+        catch (ArgumentException ex)
         {
-            ModelState.AddModelError(nameof(createSprintDto.EndDate), "End date must be after start date.");
+            ModelState.AddModelError(nameof(createSprintDto.EndDate), ex.Message);
             return BadRequest(ModelState);
         }
-
-        var sprint = new Sprint
-        {
-            Name = createSprintDto.Name,
-            StartDate = createSprintDto.StartDate.ToUniversalTime(),
-            EndDate = createSprintDto.EndDate.ToUniversalTime(),
-            TeamId = createSprintDto.TeamId
-        };
-
-        _context.Sprints.Add(sprint);
-        await _context.SaveChangesAsync();
-
-        var sprintDto = new SprintDto
-        {
-            Id = sprint.Id,
-            Name = sprint.Name,
-            StartDate = sprint.StartDate,
-            EndDate = sprint.EndDate,
-            TeamId = sprint.TeamId
-        };
-
-        return CreatedAtAction(nameof(GetSprint), new { sprintId = sprint.Id }, sprintDto);
     }
 
     /// <summary>
     /// Updates a sprint.
-    /// Only the team's admin or a super-admin can update a sprint.
     /// </summary>
-    /// <param name="sprintId">The ID of the sprint to update.</param>
-    /// <param name="updateSprintDto">The new data for the sprint.</param>
-    /// <returns>NoContent if successful, or an error response.</returns>
     [HttpPut("{sprintId}")]
     [Authorize(Policy = "IsTeamAdmin")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> UpdateSprint(Guid sprintId, UpdateSprintDto updateSprintDto)
     {
-        var sprint = await _context.Sprints
-            .Include(s => s.MoodEntries)
-            .FirstOrDefaultAsync(s => s.Id == sprintId);
-
-        if (sprint == null)
+        try
+        {
+            await _sprintService.UpdateSprintAsync(sprintId, updateSprintDto);
+            return NoContent();
+        }
+        catch (KeyNotFoundException)
         {
             return NotFound();
         }
-
-        if (updateSprintDto.EndDate <= updateSprintDto.StartDate)
+        catch (ArgumentException ex)
         {
-            ModelState.AddModelError(nameof(updateSprintDto.EndDate), "End date must be after start date.");
+            ModelState.AddModelError(nameof(updateSprintDto.EndDate), ex.Message);
             return BadRequest(ModelState);
         }
-
-        var newStart = updateSprintDto.StartDate.ToUniversalTime();
-        var newEnd = updateSprintDto.EndDate.ToUniversalTime();
-
-        // Validate date range against existing mood entries
-        if (sprint.MoodEntries.Any())
+        catch (InvalidOperationException ex)
         {
-            var minMoodDate = sprint.MoodEntries.Min(m => m.Date);
-            var maxMoodDate = sprint.MoodEntries.Max(m => m.Date);
-
-            if (newStart > minMoodDate)
+            // Map specific validation errors to ModelState for consistency with original controller
+            if (ex.Message.Contains("start date"))
             {
-                ModelState.AddModelError(nameof(updateSprintDto.StartDate), $"Cannot set start date after {minMoodDate:yyyy-MM-dd} because there are existing mood entries.");
+                ModelState.AddModelError(nameof(updateSprintDto.StartDate), ex.Message);
             }
-
-            if (newEnd < maxMoodDate)
+            else if (ex.Message.Contains("end date"))
             {
-                ModelState.AddModelError(nameof(updateSprintDto.EndDate), $"Cannot set end date before {maxMoodDate:yyyy-MM-dd} because there are existing mood entries.");
+                ModelState.AddModelError(nameof(updateSprintDto.EndDate), ex.Message);
             }
-
-            if (!ModelState.IsValid)
+            else
             {
-                return BadRequest(ModelState);
+                return BadRequest(ex.Message);
             }
+            return BadRequest(ModelState);
         }
-
-        sprint.Name = updateSprintDto.Name;
-        sprint.StartDate = newStart;
-        sprint.EndDate = newEnd;
-
-        await _context.SaveChangesAsync();
-
-        return NoContent();
     }
 
     /// <summary>
     /// Deletes a sprint.
-    /// Only the team's admin or a super-admin can delete a sprint.
     /// </summary>
-    /// <param name="sprintId">The ID of the sprint to delete.</param>
-    /// <returns>NoContent if successful, or an error response.</returns>
     [HttpDelete("{sprintId}")]
     [Authorize(Policy = "IsTeamAdmin")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> DeleteSprint(Guid sprintId)
     {
-        var sprint = await _context.Sprints.Include(s => s.Team).FirstOrDefaultAsync(s => s.Id == sprintId);
-        if (sprint == null)
+        try
+        {
+            await _sprintService.DeleteSprintAsync(sprintId);
+            return NoContent();
+        }
+        catch (KeyNotFoundException)
         {
             return NotFound();
         }
-
-        _context.Sprints.Remove(sprint);
-        await _context.SaveChangesAsync();
-
-        return NoContent();
     }
 }
