+++ api/NikoNiko.Core/Interfaces/IMoodService.cs
@@ -8,6 +8,6 @@ public interface IMoodService
     Task<IEnumerable<MoodEntryDto>> GetMoodEntriesAsync(Guid userId, bool isSuperAdmin);
     Task<PagedResult<MoodEntryDto>> GetMyMoodEntriesAsync(Guid userId, int page, int pageSize);
     Task<MoodEntryDto?> GetMoodEntryByIdAsync(Guid moodEntryId);
-    Task<MoodEntryDto> CreateOrUpdateMoodEntryAsync(CreateMoodEntryDto createMoodEntryDto, Guid authenticatedUserId);
+    Task<(MoodEntryDto dto, bool isCreated)> CreateOrUpdateMoodEntryAsync(CreateMoodEntryDto createMoodEntryDto, Guid authenticatedUserId);
     Task<IEnumerable<MoodEntryDto>> GetMoodEntriesBySprintAsync(Guid sprintId, Guid? userId, DateTime? date, Guid authenticatedUserId, bool isSuperAdmin);
 }
